Reorganise code gen helper in roc_mono
